JavaScript/Express fix
