+++ frameworks/JavaScript/express/.gitignore
@@ -1,0 +1,2 @@
+package-lock.json
+node_modules
