+++ frameworks/JavaScript/express/setup-mysql.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
-fw_depends mysql
+fw_depends nodejs mysql
 
-source ./setup.sh
+# install dependencies
+npm install
+# run app
+NODE_ENV=production node mysql-app &
